requirements.txt
boto3==1.18.6
requests==2.25.1
grpcio-tools==1.39.0
PyJWT==2.1.0
cryptography==3.4.7
